Add optimized logf
This implementation is based on code from ARM optimized routines [1].
The error for expf, for nearest rouding mode, is estimated on 0.82 ulp
with worst case.

It uses internal double precision arithmetics and a lookup table to
allow smaller polynomial and avoid the use of division.   The const
data is kept in a separate translation unit that allows fast code
for literal loads on most targets.

The performance shows about 1.7x better throughput and 1.2 better
latency for aarch64 (A72) using logf benchmark.

[1] https://github.com/ARM-software/optimized-routines

Test: ran 32-bit and 64-bit x86 tests on host
Change-Id: I5ac0b6337f851a32c37b6c0e672f0215e9cabbff